# Fix authority stat queries when a --from flag is used